+++ app/src/main/java/net/gsantner/opoc/util/ContextUtils.java
@@ -1,26 +1,23 @@
-/*
- * ------------------------------------------------------------------------------
- * Gregor Santner <gsantner.net> wrote this. You can do whatever you want
- * with it. If we meet some day, and you think it is worth it, you can buy me a
- * coke in return. Provided as is without any kind of warranty. Do not blame or
- * sue me if something goes wrong. No attribution required.    - Gregor Santner
+/*#######################################################
  *
- * License: Creative Commons Zero (CC0 1.0)
- *  http://creativecommons.org/publicdomain/zero/1.0/
- * ----------------------------------------------------------------------------
- */
+ *   Maintained by Gregor Santner, 2016-
+ *   https://gsantner.net/
+ *
+ *   License: Apache 2.0
+ *  https://github.com/gsantner/opoc/#licensing
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+#########################################################*/
 package net.gsantner.opoc.util;
 
 import android.annotation.SuppressLint;
 import android.app.AlarmManager;
 import android.app.PendingIntent;
 import android.content.ActivityNotFoundException;
-import android.content.ClipData;
 import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
-import android.content.res.ColorStateList;
 import android.content.res.Configuration;
 import android.content.res.Resources;
 import android.graphics.Bitmap;
@@ -30,9 +27,11 @@ import android.graphics.Color;
 import android.graphics.Matrix;
 import android.graphics.Paint;
 import android.graphics.Rect;
+import android.graphics.drawable.AdaptiveIconDrawable;
 import android.graphics.drawable.BitmapDrawable;
 import android.graphics.drawable.Drawable;
 import android.graphics.drawable.VectorDrawable;
+import android.media.MediaScannerConnection;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
 import android.net.Uri;
@@ -46,19 +45,17 @@ import android.support.annotation.StringRes;
 import android.support.graphics.drawable.VectorDrawableCompat;
 import android.support.v4.content.ContextCompat;
 import android.support.v4.graphics.drawable.DrawableCompat;
-import android.support.v7.widget.AppCompatButton;
 import android.text.Html;
 import android.text.SpannableString;
 import android.text.Spanned;
 import android.text.TextUtils;
 import android.text.method.LinkMovementMethod;
 import android.util.DisplayMetrics;
+import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import net.gsantner.opoc.format.markdown.SimpleMarkdownParser;
 
@@ -71,13 +68,14 @@ import java.io.InputStreamReader;
 import java.lang.reflect.Method;
 import java.util.Locale;
 
+import static android.content.Intent.FLAG_ACTIVITY_NEW_TASK;
 import static android.graphics.Bitmap.CompressFormat;
 
-@SuppressWarnings({"WeakerAccess", "unused", "SameParameterValue", "SpellCheckingInspection", "deprecation", "ObsoleteSdkInt", "ConstantConditions", "UnusedReturnValue"})
+@SuppressWarnings({"WeakerAccess", "unused", "SameParameterValue", "ObsoleteSdkInt", "deprecation", "SpellCheckingInspection"})
 public class ContextUtils {
-    //########################
-    //## Members, Constructors
-    //########################
+    //
+    // Members, Constructors
+    //
     protected Context _context;
 
     public ContextUtils(Context context) {
@@ -88,52 +86,81 @@ public class ContextUtils {
         return _context;
     }
 
-    //########################
-    //##    Resources
-    //########################
-    static class ResType {
-        public static final String DRAWABLE = "drawable";
-        public static final String STRING = "string";
-        public static final String PLURAL = "plural";
-        public static final String COLOR = "color";
-        public static final String STYLE = "style";
-        public static final String ARRAY = "array";
-        public static final String DIMEN = "dimen";
-        public static final String MENU = "menu";
-        public static final String RAW = "raw";
+
+    //
+    // Class Methods
+    //
+    public enum ResType {
+        ID, BOOL, INTEGER, COLOR, STRING, ARRAY, DRAWABLE, PLURALS,
+        ANIM, ATTR, DIMEN, LAYOUT, MENU, RAW, STYLE, XML,
     }
 
-    public String str(@StringRes int strResId) {
+    /**
+     * Find out the nuermical ressource id by given {@link ResType}
+     *
+     * @return A valid id if the id could be found, else 0
+     */
+    public int getResId(ResType resType, final String name) {
+        return _context.getResources().getIdentifier(name, resType.name().toLowerCase(), _context.getPackageName());
+    }
+
+    /**
+     * Get String by given string ressource id (nuermic)
+     */
+    public String rstr(@StringRes int strResId) {
         return _context.getString(strResId);
     }
 
-    public Drawable drawable(@DrawableRes int resId) {
+    /**
+     * Get String by given string ressource identifier (textual)
+     */
+    public String rstr(String strResKey) {
+        try {
+            return rstr(getResId(ResType.STRING, strResKey));
+        } catch (Resources.NotFoundException e) {
+            return null;
+        }
+    }
+
+    /**
+     * Get drawable from given ressource identifier
+     */
+    public Drawable rdrawable(@DrawableRes int resId) {
         return ContextCompat.getDrawable(_context, resId);
     }
 
-    public int color(@ColorRes int resId) {
+    /**
+     * Get color by given color ressource id
+     */
+    public int rcolor(@ColorRes int resId) {
         return ContextCompat.getColor(_context, resId);
     }
 
-    public int getResId(final String type, final String name) {
-        return _context.getResources().getIdentifier(name, type, _context.getPackageName());
-    }
-
-    public boolean areResIdsAvailable(final String type, final String... names) {
-        for (String name : names) {
-            if (getResId(type, name) == 0) {
+    /**
+     * Checks if all given (textual) ressource ids are available
+     *
+     * @param resType       A {@link ResType}
+     * @param resIdsTextual A (textual) identifier to be awaited at R.restype.resIdsTextual
+     * @return True if all given ids are available
+     */
+    public boolean areRessourcesAvailable(final ResType resType, final String... resIdsTextual) {
+        for (String name : resIdsTextual) {
+            if (getResId(resType, name) == 0) {
                 return false;
             }
         }
         return true;
     }
 
-    //########################
-    //##    Methods
-    //########################
-
-    public String colorToHexString(int intColor) {
-        return String.format("#%06X", 0xFFFFFF & intColor);
+    /**
+     * Convert an int color to a hex string. Optionally including alpha value.
+     *
+     * @param intColor  The color coded in int
+     * @param withAlpha Optional; Set first bool parameter to true to also include alpha value
+     */
+    public String colorToHexString(int intColor, boolean... withAlpha) {
+        boolean a = withAlpha != null && withAlpha.length >= 1 && withAlpha[0];
+        return String.format(a ? "#%08X" : "#%06X", (a ? 0xFFFFFFFF : 0xFFFFFF) & intColor);
     }
 
     public String getAppVersionName() {
@@ -147,22 +174,34 @@ public class ContextUtils {
         }
     }
 
+    /**
+     * Send a {@link Intent#ACTION_VIEW} Intent with given paramter
+     * If the parameter is an string a browser will get triggered
+     */
     public void openWebpageInExternalBrowser(final String url) {
         Uri uri = Uri.parse(url);
         Intent intent = new Intent(Intent.ACTION_VIEW, uri);
-        intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-        _context.startActivity(intent);
+        intent.addFlags(FLAG_ACTIVITY_NEW_TASK);
+        try {
+            _context.startActivity(intent);
+        } catch (ActivityNotFoundException e) {
+            e.printStackTrace();
+        }
     }
 
     /**
-     * Get field from PackageId.BuildConfig
+     * Get field from ${applicationId}.BuildConfig
      * May be helpful in libraries, where a access to
      * BuildConfig would only get values of the library
-     * rather than the app ones
+     * rather than the app ones. It awaits a string resource
+     * of the package set in manifest (root element).
+     * Falls back to applicationId of the app which may differ from manifest.
      */
     public Object getBuildConfigValue(String fieldName) {
+        String pkg = rstr("manifest_package_id");
+        pkg = (pkg != null ? pkg : _context.getPackageName()) + ".BuildConfig";
         try {
-            Class<?> c = Class.forName(_context.getPackageName() + ".BuildConfig");
+            Class<?> c = Class.forName(pkg);
             return c.getField(fieldName).get(null);
         } catch (Exception e) {
             e.printStackTrace();
@@ -170,7 +209,10 @@ public class ContextUtils {
         }
     }
 
-    public boolean getBuildConfigBoolean(String fieldName, boolean defaultValue) {
+    /**
+     * Get a BuildConfig bool value
+     */
+    public Boolean bcbool(String fieldName, Boolean defaultValue) {
         Object field = getBuildConfigValue(fieldName);
         if (field != null && field instanceof Boolean) {
             return (Boolean) field;
@@ -178,27 +220,47 @@ public class ContextUtils {
         return defaultValue;
     }
 
+    /**
+     * Get a BuildConfig string value
+     */
+    public String bcstr(String fieldName, String defaultValue) {
+        Object field = getBuildConfigValue(fieldName);
+        if (field != null && field instanceof String) {
+            return (String) field;
+        }
+        return defaultValue;
+    }
+
+    /**
+     * Check if this is a gplay build (requires BuildConfig field)
+     */
     public boolean isGooglePlayBuild() {
-        return getBuildConfigBoolean("IS_GPLAY_BUILD", true);
+        return bcbool("IS_GPLAY_BUILD", true);
     }
 
+    /**
+     * Check if this is a foss build (requires BuildConfig field)
+     */
     public boolean isFossBuild() {
-        return getBuildConfigBoolean("IS_FOSS_BUILD", false);
+        return bcbool("IS_FOSS_BUILD", false);
     }
 
-    // Requires donate__bitcoin_* resources (see below) to be available as string resource
-    public void showDonateBitcoinRequest(@StringRes final int strResBitcoinId, @StringRes final int strResBitcoinAmount, @StringRes final int strResBitcoinMessage, @StringRes final int strResAlternativeDonateUrl) {
+    /**
+     * Request a bitcoin donation with given details.
+     * All parameters are awaited as string resource ids
+     */
+    public void showDonateBitcoinRequest(@StringRes final int srBitcoinId, @StringRes final int srBitcoinAmount, @StringRes final int srBitcoinMessage, @StringRes final int srAlternativeDonateUrl) {
         if (!isGooglePlayBuild()) {
             String btcUri = String.format("bitcoin:%s?amount=%s&label=%s&message=%s",
-                    str(strResBitcoinId), str(strResBitcoinAmount),
-                    str(strResBitcoinMessage), str(strResBitcoinMessage));
+                    rstr(srBitcoinId), rstr(srBitcoinAmount),
+                    rstr(srBitcoinMessage), rstr(srBitcoinMessage));
             Intent intent = new Intent(Intent.ACTION_VIEW);
             intent.setData(Uri.parse(btcUri));
-            intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+            intent.addFlags(FLAG_ACTIVITY_NEW_TASK);
             try {
                 _context.startActivity(intent);
             } catch (ActivityNotFoundException e) {
-                openWebpageInExternalBrowser(str(strResAlternativeDonateUrl));
+                openWebpageInExternalBrowser(rstr(srAlternativeDonateUrl));
             }
         }
     }
@@ -231,44 +293,61 @@ public class ContextUtils {
         return sb.toString();
     }
 
-    @SuppressLint("RestrictedApi")
-    @SuppressWarnings("RestrictedApi")
-    public void setTintColorOfButton(AppCompatButton button, @ColorRes int resColor) {
-        button.setSupportBackgroundTintList(ColorStateList.valueOf(
-                color(resColor)
-        ));
-    }
-
-    @SuppressLint("MissingPermission") // ACCESS_NETWORK_STATE required
+    /**
+     * Get internet connection state - the permission ACCESS_NETWORK_STATE is required
+     *
+     * @return True if internet connection available
+     */
     public boolean isConnectedToInternet() {
-        ConnectivityManager con = (ConnectivityManager) _context.getSystemService(Context.CONNECTIVITY_SERVICE);
-        NetworkInfo activeNetInfo = con == null ? null : con.getActiveNetworkInfo();
-        return activeNetInfo != null && activeNetInfo.isConnectedOrConnecting();
+        try {
+            ConnectivityManager con = (ConnectivityManager) _context.getSystemService(Context.CONNECTIVITY_SERVICE);
+            @SuppressLint("MissingPermission") NetworkInfo activeNetInfo =
+                    con == null ? null : con.getActiveNetworkInfo();
+            return activeNetInfo != null && activeNetInfo.isConnectedOrConnecting();
+        } catch (Exception ignored) {
+            throw new RuntimeException("Error: Developer forgot to declare a permission");
+        }
     }
 
-    public boolean isConnectedToInternet(@Nullable @StringRes Integer warnMessageStringRes) {
-        final boolean result = isConnectedToInternet();
-        if (!result && warnMessageStringRes != null)
-            Toast.makeText(_context, _context.getString(warnMessageStringRes), Toast.LENGTH_SHORT).show();
-
-        return result;
+    /**
+     * Check if app with given {@code packageName} is installed
+     */
+    public boolean isAppInstalled(String packageName) {
+        PackageManager pm = _context.getApplicationContext().getPackageManager();
+        try {
+            pm.getPackageInfo(packageName, PackageManager.GET_ACTIVITIES);
+            return true;
+        } catch (PackageManager.NameNotFoundException e) {
+            return false;
+        }
     }
 
-    public void restartApp(Class classToStartupWith) {
-        Intent restartIntent = new Intent(_context, classToStartupWith);
-        PendingIntent restartIntentP = PendingIntent.getActivity(_context, 555,
-                restartIntent, PendingIntent.FLAG_CANCEL_CURRENT);
+    /**
+     * Restart the current app. Supply the class to start on startup
+     */
+    public void restartApp(Class classToStart) {
+        Intent inte = new Intent(_context, classToStart);
+        PendingIntent inteP = PendingIntent.getActivity(_context, 555, inte, PendingIntent.FLAG_CANCEL_CURRENT);
         AlarmManager mgr = (AlarmManager) _context.getSystemService(Context.ALARM_SERVICE);
-        mgr.set(AlarmManager.RTC, System.currentTimeMillis() + 100, restartIntentP);
-        System.exit(0);
+        if (mgr != null) {
+            mgr.set(AlarmManager.RTC, System.currentTimeMillis() + 100, inteP);
+        } else {
+            inte.addFlags(FLAG_ACTIVITY_NEW_TASK);
+            _context.startActivity(inte);
+        }
+        Runtime.getRuntime().exit(0);
     }
 
+    /**
+     * Load a markdown file from a {@link RawRes}, prepend each line with {@code prepend} text
+     * and convert markdown to html using {@link SimpleMarkdownParser}
+     */
     public String loadMarkdownForTextViewFromRaw(@RawRes int rawMdFile, String prepend) {
         try {
             return new SimpleMarkdownParser()
                     .parse(_context.getResources().openRawResource(rawMdFile),
                             prepend, SimpleMarkdownParser.FILTER_ANDROID_TEXTVIEW)
-                    .replaceColor("#000001", color(getResId(ResType.COLOR, "accent")))
+                    .replaceColor("#000001", rcolor(getResId(ResType.COLOR, "accent")))
                     .removeMultiNewlines().replaceBulletCharacter("*").getHtml();
         } catch (IOException e) {
             e.printStackTrace();
@@ -276,53 +355,74 @@ public class ContextUtils {
         }
     }
 
+    /**
+     * Load html into a {@link Spanned} object and set the
+     * {@link TextView}'s text using {@link TextView#setText(CharSequence)}
+     */
     public void setHtmlToTextView(TextView textView, String html) {
         textView.setMovementMethod(LinkMovementMethod.getInstance());
         textView.setText(new SpannableString(htmlToSpanned(html)));
     }
 
+    /**
+     * Estimate this device's screen diagonal size in inches
+     */
     public double getEstimatedScreenSizeInches() {
         DisplayMetrics dm = _context.getResources().getDisplayMetrics();
 
-        double density = dm.density * 160;
-        double x = Math.pow(dm.widthPixels / density, 2);
-        double y = Math.pow(dm.heightPixels / density, 2);
-        double screenInches = Math.sqrt(x + y) * 1.16;  // 1.16 = est. Nav/Statusbar
-        screenInches = screenInches < 4.0 ? 4.0 : screenInches;
-        screenInches = screenInches > 12.0 ? 12.0 : screenInches;
-        return screenInches;
+        double calc = dm.density * 160d;
+        double x = Math.pow(dm.widthPixels / calc, 2);
+        double y = Math.pow(dm.heightPixels / calc, 2);
+        calc = Math.sqrt(x + y) * 1.16;  // 1.16 = est. Nav/Statusbar
+        return Math.min(12, Math.max(4, calc));
     }
 
+    /**
+     * Check if the device is currently in portrait orientation
+     */
     public boolean isInPortraitMode() {
         return _context.getResources().getConfiguration().orientation == Configuration.ORIENTATION_PORTRAIT;
     }
 
-    public Locale getLocaleByAndroidCode(String code) {
-        if (!TextUtils.isEmpty(code)) {
-            return code.contains("-r")
-                    ? new Locale(code.substring(0, 2), code.substring(4, 6)) // de-rAt
-                    : new Locale(code); // de
+    /**
+     * Get an {@link Locale} out of a android language code
+     * The {@code androidLC} may be in any of the forms: de, en, de-rAt
+     */
+    public Locale getLocaleByAndroidCode(String androidLC) {
+        if (!TextUtils.isEmpty(androidLC)) {
+            return androidLC.contains("-r")
+                    ? new Locale(androidLC.substring(0, 2), androidLC.substring(4, 6)) // de-rAt
+                    : new Locale(androidLC); // de
         }
         return Resources.getSystem().getConfiguration().locale;
     }
 
-    //  en/de/de-rAt ; Empty string -> default locale
-    public void setAppLanguage(String androidLocaleString) {
-        Locale locale = getLocaleByAndroidCode(androidLocaleString);
+    /**
+     * Set the apps language
+     * {@code androidLC} may be in any of the forms: en, de, de-rAt
+     * If given an empty string, the default (system) locale gets loaded
+     */
+    public void setAppLanguage(String androidLC) {
+        Locale locale = getLocaleByAndroidCode(androidLC);
         Configuration config = _context.getResources().getConfiguration();
-        config.locale = (locale != null && !androidLocaleString.isEmpty())
+        config.locale = (locale != null && !androidLC.isEmpty())
                 ? locale : Resources.getSystem().getConfiguration().locale;
         _context.getResources().updateConfiguration(config, null);
     }
 
-    // Find out if color above the given color should be light or dark. true if light
+    /**
+     * Try to guess if the color on top of the given {@code colorOnBottomInt}
+     * should be light or dark. Returns true if top color should be light
+     */
     public boolean shouldColorOnTopBeLight(@ColorInt int colorOnBottomInt) {
         return 186 > (((0.299 * Color.red(colorOnBottomInt))
                 + ((0.587 * Color.green(colorOnBottomInt))
                 + (0.114 * Color.blue(colorOnBottomInt)))));
     }
 
-    @SuppressWarnings("deprecation")
+    /**
+     * Convert a html string to an android {@link Spanned} object
+     */
     public Spanned htmlToSpanned(String html) {
         Spanned result;
         if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.N) {
@@ -333,49 +433,56 @@ public class ContextUtils {
         return result;
     }
 
-    public void setClipboard(String text) {
-        if (android.os.Build.VERSION.SDK_INT < android.os.Build.VERSION_CODES.HONEYCOMB) {
-            ((android.text.ClipboardManager) _context.getSystemService(Context.CLIPBOARD_SERVICE)).setText(text);
-        } else {
-            ClipData clip = ClipData.newPlainText(_context.getPackageName(), text);
-            ((android.content.ClipboardManager) _context.getSystemService(Context.CLIPBOARD_SERVICE)).setPrimaryClip(clip);
-        }
-    }
-
-    public String[] getClipboard() {
-        String[] ret;
-        if (android.os.Build.VERSION.SDK_INT < android.os.Build.VERSION_CODES.HONEYCOMB) {
-            ret = new String[]{((android.text.ClipboardManager) _context.getSystemService(Context.CLIPBOARD_SERVICE)).getText().toString()};
-        } else {
-            ClipData data = ((android.content.ClipboardManager) _context.getSystemService(Context.CLIPBOARD_SERVICE)).getPrimaryClip();
-            ret = new String[data.getItemCount()];
-            for (int i = 0; i < data.getItemCount() && i < ret.length; i++) {
-                ret[i] = data.getItemAt(i).getText().toString();
-            }
-        }
-        return ret;
-    }
-
-    public float px2dp(final float px) {
+    /**
+     * Convert pixel unit do android dp unit
+     */
+    public float convertPxToDp(final float px) {
         return px / _context.getResources().getDisplayMetrics().density;
     }
 
-    public float dp2px(final float dp) {
+    /**
+     * Convert android dp unit to pixel unit
+     */
+    public float convertDpToPx(final float dp) {
         return dp * _context.getResources().getDisplayMetrics().density;
     }
 
-    public void setViewVisible(View view, boolean visible) {
-        view.setVisibility(visible ? View.VISIBLE : View.GONE);
+    /**
+     * Request the givens paths to be scanned by MediaScanner
+     *
+     * @param files Files and folders to scan
+     */
+    public void mediaScannerScanFile(File... files) {
+        if (android.os.Build.VERSION.SDK_INT > 19) {
+            String[] paths = new String[files.length];
+            for (int i = 0; i < files.length; i++) {
+                paths[i] = files[i].getAbsolutePath();
+            }
+            MediaScannerConnection.scanFile(_context, paths, null, null);
+        } else {
+            for (File file : files) {
+                _context.sendBroadcast(new Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE, Uri.fromFile(file)));
+            }
+        }
     }
 
+    /**
+     * Load an image into a {@link ImageView} and apply a color filter
+     */
     public static void setDrawableWithColorToImageView(ImageView imageView, @DrawableRes int drawableResId, @ColorRes int colorResId) {
         imageView.setImageResource(drawableResId);
         imageView.setColorFilter(ContextCompat.getColor(imageView.getContext(), colorResId));
     }
 
+    /**
+     * Get a {@link Bitmap} out of a {@link Drawable}
+     */
     public Bitmap drawableToBitmap(Drawable drawable) {
         Bitmap bitmap = null;
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP && (drawable instanceof VectorDrawable || drawable instanceof VectorDrawableCompat)) {
+        if (drawable instanceof VectorDrawableCompat
+                || (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP && drawable instanceof VectorDrawable)
+                || ((Build.VERSION.SDK_INT >= Build.VERSION_CODES.O && drawable instanceof AdaptiveIconDrawable))) {
+
             if (Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP) {
                 drawable = (DrawableCompat.wrap(drawable)).mutate();
             }
@@ -391,6 +498,18 @@ public class ContextUtils {
         return bitmap;
     }
 
+    /**
+     * Get a {@link Bitmap} out of a {@link DrawableRes}
+     */
+    public Bitmap drawableToBitmap(@DrawableRes int drawableId) {
+        return drawableToBitmap(ContextCompat.getDrawable(_context, drawableId));
+    }
+
+    /**
+     * Get a {@link Bitmap} from a given {@code imagePath} on the filesystem
+     * Specifying a {@code maxDimen} is also possible and a value below 2000
+     * is recommended, otherwise a {@link OutOfMemoryError} may occur
+     */
     public Bitmap loadImageFromFilesystem(File imagePath, int maxDimen) {
         BitmapFactory.Options options = new BitmapFactory.Options();
         options.inJustDecodeBounds = true;
@@ -419,6 +538,10 @@ public class ContextUtils {
         return inSampleSize;
     }
 
+    /**
+     * Scale the bitmap so both dimensions are lower or equal to {@code maxDimen}
+     * This keeps the aspect ratio
+     */
     public Bitmap scaleBitmap(Bitmap bitmap, int maxDimen) {
         int picSize = Math.min(bitmap.getHeight(), bitmap.getWidth());
         float scale = 1.f * maxDimen / picSize;
@@ -427,31 +550,43 @@ public class ContextUtils {
         return Bitmap.createBitmap(bitmap, 0, 0, bitmap.getWidth(), bitmap.getHeight(), matrix, true);
     }
 
-    public File writeImageToFileJpeg(File imageFile, Bitmap image) {
+    /**
+     * Write the given {@link Bitmap} to {@code imageFile}, in {@link CompressFormat#JPEG} format
+     */
+    public boolean writeImageToFileJpeg(File imageFile, Bitmap image) {
         return writeImageToFile(imageFile, image, Bitmap.CompressFormat.JPEG, 95);
     }
 
-
-    public File writeImageToFileDetectFormat(File imageFile, Bitmap image, int quality) {
-        CompressFormat format = CompressFormat.JPEG;
-        String lc = imageFile.getAbsolutePath().toLowerCase(Locale.ROOT);
-        if (lc.endsWith(".png")) {
-            format = CompressFormat.PNG;
+    /**
+     * Write the given {@link Bitmap} to filesystem
+     *
+     * @param targetFile The file to be written in
+     * @param image      The image as android {@link Bitmap}
+     * @param format     One format of {@link CompressFormat}, null will determine based on filename
+     * @param quality    Quality level, defaults to 95
+     * @return True if writing was successful
+     */
+    public boolean writeImageToFile(File targetFile, Bitmap image, CompressFormat format, Integer quality) {
+        File folder = new File(targetFile.getParent());
+        if (quality == null || quality < 0 || quality > 100) {
+            quality = 95;
         }
-        if (lc.endsWith(".webp")) {
-            format = CompressFormat.WEBP;
+        if (format == null) {
+            format = CompressFormat.JPEG;
+            String lc = targetFile.getAbsolutePath().toLowerCase(Locale.ROOT);
+            if (lc.endsWith(".png")) {
+                format = CompressFormat.PNG;
+            }
+            if (lc.endsWith(".webp")) {
+                format = CompressFormat.WEBP;
+            }
         }
-        return writeImageToFile(imageFile, image, format, quality);
-    }
-
-    public File writeImageToFile(File imageFile, Bitmap image, CompressFormat format, int quality) {
-        File folder = new File(imageFile.getParent());
         if (folder.exists() || folder.mkdirs()) {
             FileOutputStream stream = null;
             try {
-                stream = new FileOutputStream(imageFile); // overwrites this image every time
+                stream = new FileOutputStream(targetFile); // overwrites this image every time
                 image.compress(format, quality, stream);
-                return imageFile;
+                return true;
             } catch (FileNotFoundException ignored) {
             } finally {
                 try {
@@ -462,13 +597,17 @@ public class ContextUtils {
                 }
             }
         }
-        return null;
+        return false;
     }
 
-    public Bitmap drawTextToDrawable(@DrawableRes int resId, String text, int textSize) {
+    /**
+     * Draw text in the center of the given {@link DrawableRes}
+     * This may be useful for e.g. badge counts
+     */
+    public Bitmap drawTextOnDrawable(@DrawableRes int drawableRes, String text, int textSize) {
         Resources resources = _context.getResources();
         float scale = resources.getDisplayMetrics().density;
-        Bitmap bitmap = getBitmapFromDrawable(resId);
+        Bitmap bitmap = drawableToBitmap(drawableRes);
 
         bitmap = bitmap.copy(bitmap.getConfig(), true);
         Canvas canvas = new Canvas(bitmap);
@@ -486,26 +625,11 @@ public class ContextUtils {
         return bitmap;
     }
 
-    public Bitmap getBitmapFromDrawable(int drawableId) {
-        Bitmap bitmap = null;
-        Drawable drawable = ContextCompat.getDrawable(_context, drawableId);
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP && (drawable instanceof VectorDrawable || drawable instanceof VectorDrawableCompat)) {
-            if (Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP) {
-                drawable = (DrawableCompat.wrap(drawable)).mutate();
-            }
-
-            bitmap = Bitmap.createBitmap(drawable.getIntrinsicWidth(),
-                    drawable.getIntrinsicHeight(), Bitmap.Config.ARGB_8888);
-            Canvas canvas = new Canvas(bitmap);
-            drawable.setBounds(0, 0, canvas.getWidth(), canvas.getHeight());
-            drawable.draw(canvas);
-        } else if (drawable instanceof BitmapDrawable) {
-            bitmap = ((BitmapDrawable) drawable).getBitmap();
-        }
-        return bitmap;
-    }
-
-    public ContextUtils tintMenuItems(Menu menu, boolean recurse, @ColorInt int iconColor) {
+    /**
+     * Try to tint all {@link Menu}s {@link MenuItem}s with given color
+     */
+    @SuppressWarnings("ConstantConditions")
+    public void tintMenuItems(Menu menu, boolean recurse, @ColorInt int iconColor) {
         for (int i = 0; i < menu.size(); i++) {
             MenuItem item = menu.getItem(i);
             tintDrawable(item.getIcon(), iconColor);
@@ -513,13 +637,18 @@ public class ContextUtils {
                 tintMenuItems(item.getSubMenu(), recurse, iconColor);
             }
         }
-        return this;
     }
 
+    /**
+     * Loads {@link Drawable} by given {@link DrawableRes} and applies a color
+     */
     public Drawable tintDrawable(@DrawableRes int drawableRes, @ColorInt int color) {
-        return tintDrawable(_context.getResources().getDrawable(drawableRes), color);
+        return tintDrawable(rdrawable(drawableRes), color);
     }
 
+    /**
+     * Tint a {@link Drawable} with given {@code color}
+     */
     public Drawable tintDrawable(@Nullable Drawable drawable, @ColorInt int color) {
         if (drawable != null) {
             drawable = DrawableCompat.wrap(drawable);
@@ -528,16 +657,19 @@ public class ContextUtils {
         return drawable;
     }
 
-    @SuppressLint("PrivateApi")
-    public ContextUtils setSubMenuIconsVisiblity(Menu menu, boolean visible) {
+    /**
+     * Try to make icons in Toolbar/ActionBars SubMenus visible
+     * This may not work on some devices and it maybe won't work on future android updates
+     */
+    public void setSubMenuIconsVisiblity(Menu menu, boolean visible) {
         if (menu.getClass().getSimpleName().equals("MenuBuilder")) {
             try {
-                Method m = menu.getClass().getDeclaredMethod("setOptionalIconsVisible", Boolean.TYPE);
+                @SuppressLint("PrivateApi") Method m = menu.getClass().getDeclaredMethod("setOptionalIconsVisible", Boolean.TYPE);
                 m.setAccessible(true);
                 m.invoke(menu, visible);
             } catch (Exception ignored) {
+                Log.d(getClass().getName(), "Error: 'setSubMenuIconsVisiblity' not supported on this device");
             }
         }
-        return this;
     }
 }