+++ app/src/main/java/net/gsantner/opoc/util/NetworkUtils.java
@@ -1,15 +1,13 @@
-/*
- * ------------------------------------------------------------------------------
- * Lonami Exo <lonamiwebs.github.io> wrote this. You can do whatever you want
- * with it. If we meet some day, and you think it is worth it, you can buy me
- * a coke in return. Provided as is without any kind of warranty. Do not blame
- * or sue me if something goes wrong. No attribution required.
- *                                                             - Lonami Exo
+/*#######################################################
  *
- * License: Creative Commons Zero (CC0 1.0)
- *  http://creativecommons.org/publicdomain/zero/1.0/
- * ----------------------------------------------------------------------------
- */
+ *   Maintained by Gregor Santner, 2017-
+ *   https://gsantner.net/
+ *
+ *   License: Apache 2.0
+ *  https://github.com/gsantner/opoc/#licensing
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+#########################################################*/
 package net.gsantner.opoc.util;
 
 import org.json.JSONObject;
@@ -55,13 +53,20 @@ public class NetworkUtils {
     }
 
     public static boolean downloadFile(final URL url, final File outFile, final Callback.a1<Float> progressCallback) {
+        return downloadFile(url, outFile, null, progressCallback);
+    }
+
+    public static boolean downloadFile(final URL url, final File outFile, HttpURLConnection connection, final Callback.a1<Float> progressCallback) {
         InputStream input = null;
         OutputStream output = null;
-        HttpURLConnection connection = null;
         try {
-            connection = (HttpURLConnection) url.openConnection();
+            if (connection == null) {
+                connection = (HttpURLConnection) url.openConnection();
+            }
             connection.connect();
-            input = connection.getInputStream();
+            input = connection.getResponseCode() < HttpURLConnection.HTTP_BAD_REQUEST
+                    ? connection.getInputStream() : connection.getErrorStream();
+
 
             if (!outFile.getParentFile().isDirectory())
                 if (!outFile.getParentFile().mkdirs())
@@ -108,6 +113,15 @@ public class NetworkUtils {
         return "";
     }
 
+    public static String performCall(final String url, final String method, final String data) {
+        try {
+            return performCall(new URL(url), method, data);
+        } catch (MalformedURLException e) {
+            e.printStackTrace();
+        }
+        return "";
+    }
+
     // URL encoded parameters
     public static String performCall(final String url, final String method, final HashMap<String, String> params) {
         try {
@@ -133,20 +147,28 @@ public class NetworkUtils {
     }
 
     private static String performCall(final URL url, final String method, final String data) {
+        return performCall(url, method, data, null);
+    }
+
+    private static String performCall(final URL url, final String method, final String data, final HttpURLConnection existingConnection) {
         try {
-            final HttpURLConnection conn = (HttpURLConnection) url.openConnection();
-            conn.setRequestMethod(method);
-            conn.setDoInput(true);
+            final HttpURLConnection connection = existingConnection != null
+                    ? existingConnection : (HttpURLConnection) url.openConnection();
+            connection.setRequestMethod(method);
+            connection.setDoInput(true);
 
             if (data != null && !data.isEmpty()) {
-                conn.setDoOutput(true);
-                final OutputStream output = conn.getOutputStream();
+                connection.setDoOutput(true);
+                final OutputStream output = connection.getOutputStream();
                 output.write(data.getBytes(Charset.forName(UTF8)));
                 output.flush();
                 output.close();
             }
 
-            return FileUtils.readCloseTextStream(conn.getInputStream());
+            InputStream input = connection.getResponseCode() < HttpURLConnection.HTTP_BAD_REQUEST
+                    ? connection.getInputStream() : connection.getErrorStream();
+
+            return FileUtils.readCloseTextStream(connection.getInputStream());
         } catch (Exception e) {
             e.printStackTrace();
         }