+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### v1.4.2
+- Add some more bundled assets
+
+### v1.4.1
+- Adaptive app icon
+- Fix media scan setting option
+- Merge first-start and update dialog
+
 ### v1.3.2
 - More support languages
 - New settings screen